chore: components renamed (capitalized), btn class selector renamed (snake_case)
