+++ packages/client/src/components/Button/Button.module.scss
@@ -21,17 +21,17 @@
   cursor: default;
 }
 
-.btn-primary {
+.btn_primary {
   background: var(--color-primary);
   color: var(--color-text-on-dark);
 }
 
-.btn-secondary {
+.btn_secondary {
   background: var(--color-secondary);
   color: var(--color-text-on-dark);
 }
 
-.btn-accent {
+.btn_accent {
   background: var(--color-accent);
   color: var(--color-text-on-ligh);
 }
