+++ packages/client/src/pages/GamePage.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react'
-import { GameField } from '../components/game-field/GameField'
+import { GameField } from '../components/GameField/GameField'
 
 export const GamePage: FC = () => {
   // TODO: replace with actual game level, add game level progress, timer, etc.
